Extract participation unit tests into separate module.
